refactor: replace latexdiff thin wrappers with deep service and focused managers
